+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 
-## [Unreleased] - 2026-08-14
+## [3.1.0] - 2026-08-14
+### Added
+ - `ansi.splitAt()` function to split a string into two parts at a given visible index, ignoring
+    ANSI sequences
+
 ### Fixed
  - Cursor movement functions will no longer emit an invalid sequence containing `?` when given `0`,
    and will instead emit an empty string
@@ -195,6 +199,7 @@
  - `stripAnsiSequences()` to remove ANSI codes from a string
  - Auto-detection of color support via TTY state and `NO_COLOR`, `FORCE_COLOR`, etc.
 
+[3.1.0]: https://github.com/anderson-pete/ansi/compare/v3.0.0...v3.1.0
 [3.0.0]: https://github.com/anderson-pete/ansi/compare/v2.3.1...v3.0.0
 [2.3.1]: https://github.com/anderson-pete/ansi/compare/v2.3.0...v2.3.1
 [2.3.0]: https://github.com/anderson-pete/ansi/compare/v2.2.0...v2.3.0